Update menu and fix issues
