+++ app/src/main/java/com/parcelpaladin/d30f2n/parcelpaladinfirebase/MainActivity.java
@@ -46,12 +46,9 @@ public class MainActivity extends AppCompatActivity {
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        if(item.getItemId() == R.id.setting){
-            Toast.makeText(this, "Setting Activity", Toast.LENGTH_SHORT).show();
-        }
 
         if(item.getItemId() == R.id.logout){
-            Toast.makeText(this, "Logout Activity", Toast.LENGTH_SHORT).show();
+//            Toast.makeText(this, "Logout Activity", Toast.LENGTH_SHORT).show();
             firebaseAuth.signOut();
             finish();
             startActivity(new Intent(this, LoginActivity.class));
