+++ app/src/main/res/layout/activity_log.xml
@@ -24,18 +24,8 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.0" />
 
-    <ListView
-        android:id="@+id/failedlw"
-        android:layout_width="0dp"
-        android:layout_height="211dp"
-        android:layout_marginBottom="172dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginStart="8dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
     <TextView
+        android:id="@+id/bottom"
         android:layout_width="0dp"
         android:layout_height="38dp"
         android:layout_marginBottom="8dp"
@@ -60,8 +50,19 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/top"
 
-        />
+        app:layout_constraintTop_toBottomOf="@+id/top" />
+
+    <ListView
+        android:id="@+id/failedlw"
+        android:layout_width="0dp"
+        android:layout_height="346dp"
+        android:layout_marginEnd="8dp"
+        android:layout_marginStart="8dp"
+        app:layout_constraintTop_toBottomOf="@+id/bottom"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintStart_toStartOf="parent" />
 
 </android.support.constraint.ConstraintLayout>
