+++ app/src/main/res/menu/menu.xml
@@ -2,7 +2,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
 
-    <item android:id = "@+id/setting" android:title="Settings"></item>
     <item android:id = "@+id/logout" android:title="Logout">
